+++ server/Makefile
@@ -1,18 +1,20 @@
 TARGET_EXEC := aesdsocket
 
 BUILD_DIR := ./build
-SRC_DIR := ./
+TGT_DIR := ./
+SRC_DIR := ./src
+INC_DIR := ./include
 
 SRCS := $(wildcard $(SRC_DIR)/*.c)	# Find all c files
 OBJS := $(SRCS:$(SRC_DIR)/%.c=$(BUILD_DIR)/%.o)	# make object files for all c files 
 
 cc = ($CROSS_COMPILE)gcc
-CFLAGS = -g -Wall -Werror -O0
+CFLAGS = -g -Wall -Werror -O0 -I$(INC_DIR)
 
-all: $(SRC_DIR)/$(TARGET_EXEC)
+all: $(TGT_DIR)/$(TARGET_EXEC)
 
 # Target to build final executable
-$(SRC_DIR)/$(TARGET_EXEC): $(OBJS)
+$(TGT_DIR)/$(TARGET_EXEC): $(OBJS)
 	$(CC) $(OBJS) -o $@
 
 # Target to build the object files
@@ -26,4 +28,3 @@ clean:
 	then \
 		rm -r ${BUILD_DIR}; \
 	fi
-
